# Lock-free atomic observations in Histograms!