@mxinden 2019-03-01
Just went through this while analysing memory allocations of histograms (https://github.com/kubernetes/kubernetes/pull/74806). @beorn7 fancy, I haven't thought about the hot-cold for writes-reads before.