@beorn7 2019-03-01
Thanks. I gave a talk about it at the Berlin Gopher meetup (not recorded, but here are [slides](https://docs.google.com/presentation/d/1wuNNW-g6v8qizIc_IxAGZTj-49TODKF0TYddTA1VDUo/edit)). I was genuinely worried that somebody in the audience would tell me that there is a much easier way… But what I learned from some people in the audience is that "atomically change two separate values" is something many concurrency researchers would like to have but what they cannot get from CPU developers. Now I'm thinking I should give a refined version of that talk again at some Go conference.

Also, the code got some improvements in #536 by @pascaldekloe .